functions/review_stats_http/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.16.1

